+++ src/components/modalViews/index.ts
@@ -1,7 +1,7 @@
 export { default as ChannelSelectModal } from "./ChannelSelectModal";
 export { default as TestModal } from "./TestModal";
 export { default as TagCreateModal } from "./TagCreateModal";
-export { default as EditNameModal } from "./EditNameModal";
-export { default as EditPasswordModal } from "./EditPasswordModal";
-export { default as ChangeImageModal } from "./ChangeImageModal";
+export { default as EditNameModal } from "./ProfileModal/EditNameModal";
+export { default as EditPasswordModal } from "./ProfileModal/EditPasswordModal";
+export { default as ChangeImageModal } from "./ProfileModal/ChangeImageModal";
 export { default as PostDetailModal } from "./PostDetailModal/PostDetailModal";
